+++ src/components/atoms/Tags.astro
@@ -1,5 +1,5 @@
 ---
-import { getPermalink } from "~/utils/permalinks";
+import { getPermalink } from "~/permalink";
 
 const { tags, class: className = "text-sm" } = Astro.props;
 ---
