+++ src/components/blog/highlights/Item.astro
@@ -1,7 +1,7 @@
 ---
 import Picture from "~/components/core/Picture.astro";
 
-import {getLink} from "../../../utils/content";
+import {getLink} from "~/permalink";
 
 const { id, data } = Astro.props;
 ---
