+++ src/components/index-template.astro
@@ -1,6 +1,6 @@
 ---
 import { SITE } from "~/config.mjs";
-import { getCanonical, getHomePermalink } from "~/utils/permalinks";
+import { getCanonical, getHomePermalink } from "~/permalink";
 
 import Hero from "~/components/widgets/Hero.astro";
 import HeroBanner from "~/components/widgets/HeroBanner.astro";
