+++ src/components/widgets/Error404.astro
@@ -1,5 +1,5 @@
 ---
-import { getHomePermalink } from "~/utils/permalinks";
+import { getHomePermalink } from "~/permalink";
 
 const {} = Astro.props;
 ---
