+++ src/components/widgets/HeroBanner.astro
@@ -1,5 +1,5 @@
 ---
-import {getLink} from "../../utils/content";
+import {getLink} from "~/permalink";
 import Picture from "../core/Picture.astro";
 const {title, description, primary, secondary} = Astro.props;
