+++ src/content/pages/contact-success.en.mdx
@@ -4,10 +4,8 @@ path: contact/thanks
 variant: page
 ---
 
-import ContactForm from "../../components/widgets/Form.astro";
 import ProseSection from "../../components/widgets/ProseSection.astro";
 import Icon from "astro-iconify";
-import { getLink } from "../../utils/content.js";
 
 <div class="px-8 sm:px-6 py-12 sm:py-16 lg:py-20 mx-auto max-w-3xl">
   <ProseSection title="Contact">
