+++ src/content/pages/contact.en.mdx
@@ -6,7 +6,7 @@ variant: page
 
 import ContactForm from "../../components/widgets/Form.astro";
 import ProseSection from "../../components/widgets/ProseSection.astro";
-import { getLink } from "../../utils/content.js";
+import { getLink } from "~/permalink";
 
 <div class="px-8 sm:px-6 py-12 sm:py-16 lg:py-20 mx-auto max-w-3xl">
   <ProseSection title="Contact">
