+++ src/pages/[...locale]/blog/[...page].astro
@@ -5,7 +5,7 @@ import Layout from "~/layouts/BlogLayout.astro";
 import BlogList from "~/components/blog/List.astro";
 import Pagination from "~/components/atoms/Pagination.astro";
 
-import {getCanonical, getPermalink} from "~/utils/permalinks";
+import {getCanonical, getPermalink} from "~/permalink";
 import {postsCollection} from "../../../content";
 
 export async function getStaticPaths({paginate}) {
